+++ dynts/bench/__init__.py
@@ -1,5 +1,0 @@
-from suite import runbench
-
-import rolling
-        
-